Move repo state store files to repo-states-store dir
