+++ extensions/ql-vscode/src/variant-analysis/repo-states-store/repo-states-data-types.ts
