+++ extensions/ql-vscode/src/variant-analysis/repo-states-store/repo-states-to-domain-mapper.ts
